chore(behaviors): finish watcher->behavior rename on helper filenames and exposed-surface docs
PR #2034 purged watcher from the exposed SDK/REST/MCP/reaction surface but
left three helper files, one exported symbol, and a handful of developer-facing
doc comments on the old name:

- rename watcher-execution-config.ts -> behavior-execution-config.ts,
  watcher-device-access.ts -> behavior-device-access.ts,
  get_content/watcher-mode.ts -> behavior-mode.ts (+ their tests), matching
  the manage_watchers/ -> manage_behaviors/ rename direction of #2034
- rename handleWatcherMode -> handleBehaviorMode and fix comments that still
  claimed a watcher_id arg (the real arg is behavior_id)
- update connector-sdk reaction typedocs and sandbox client-sdk header to say
  behavior reaction, matching the exposed product surface

Internal vocabulary (watchers table, actingWatcherId, src/watchers/) is
unchanged per packages/server/AGENTS.md.
